security bugs pin fixed
